% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib.R
\name{calib}
\alias{calib}
\title{Calibrating risk-neutral parameters}
\usage{
calib(theta, price, market_price)
}
\arguments{
\item{theta}{A q times 1 vector}

\item{price}{A k times 1 vector containing the computed futures prices}

\item{market_price}{A k times 1 vector containing the actual futures prices traded in the market}
}
\value{
A list containing:
\item{}{}
\item{}{}
\item{}{}
}
\description{
Function that returns the calibrated risk-neutral parameters and in-sample pricing errors based on the market prices
}
