% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostics}
\alias{diagnostics}
\title{Descriptive analysis of deseasonalized temperatures data}
\usage{
diagnostics(resid)
}
\arguments{
\item{resid}{A n by 1 vector containing the past deseasonalized temperatures data at one station over (n / 365) years}
}
\value{
A list containing:
\item{}{}
\item{}{}
\item{}{}
}
\description{
Function used for simple preliminary tests on the deseasonalized temperatures data,
returns test statistics and p-values from tests such as the Kolmogorov-Smirnov test,
and visual outputs such as Q-Q plots and kernel density estimates
}
\examples{
# load the residuals data from residuals.rda in the data folder

# example 1
resid = as.numeric(residuals[, 3]) # deseasonalized temperaures at one station
}
