% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loc_hedge.R
\name{loc_hedge}
\alias{loc_hedge}
\title{Hedging spatial risk with optimal portfolio}
\usage{
loc_hedge(residuals, station)
}
\arguments{
\item{residuals}{A n by p matrix containing the past deseasonalized temperatures data at p stations over (n / 365) years (n >= 365 * 2) (p >= 3)}

\item{station}{A numeric index denoting the specific station to be hedged}
}
\value{
A list containing:
\item{optim_weights}{A 1 by (p - 1) vector containing the optimal weights for each station in the portfolio}
}
\description{
Function that computes and returns the optimal weights for hedging the
temperature risk at a particular location using the contracts traded based on other locations
}
\examples{
# load the residuals data from residuals.rda in the data folder

# example 1
residuals1 = matrix(as.numeric(residuals[, 3:5]), 730, 3)
station1 = 1
loc_hedge1 = loc_hedge(residuals1, station1)

# examine results
loc_hedge1$optim_weights # weights for each of the other stations in the optimal portfolio
sum(loc_hedge1$optim_weights) # sum to one

# example 2
residuals2 = matrix(as.numeric(residuals[, 51:55]), 730, 5)
station2 = 3
loc_hedge2 = loc_hedge(residuals2, station2)

# examine results
loc_hedge2$optim_weights # weights for each of the other stations in the optimal portfolio
sum(loc_hedge2$optim_weights) # sum to one
}
