% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLSS_spacetime.R
\name{MLSS_spacetime}
\alias{MLSS_spacetime}
\title{Pricing with non-Gaussian spatio-temporal random fields}
\usage{
MLSS_spacetime(residuals, station, start_ind, end_ind, type, seasonal_coefs)
}
\arguments{
\item{residuals}{A n by p matrix containing the past deseasonalized temperatures data at p stations over (n / 365) years (n >= 365 * 2) (p >= 3)}

\item{station}{A numeric index denoting the specific station that the price of the temperature futures depends on}

\item{start_ind}{A numeric index for the start of the measurement period}

\item{end_ind}{A numeric index for the end of the measurement period}

\item{type}{A string denoting the type of the temperature futures, either CDD, HDD or CAT}

\item{seasonal_coefs}{A 1 by 4 vector containing the coefficients of the seasonality function at the station to be priced}
}
\value{
A list containing:
\item{price}{a scalar that is the price of the derivative computed}
\item{mu}{1 by p vector containing the parameter mu of the fitted NIG distribution for the standardized residuals}
}
\description{
Function that returns the price of a temperature futures using non-Gaussian spatio-temporal random fields
to estimate multivariate seasonal volatility
}
\examples{
# load the residuals data from residuals.rda in the data folder
# load the seasonal coefficients data from seasonal_coefs.rda in the data folder

# example 1
residuals1 = matrix(as.numeric(residuals[, 3:5]), 730, 3)
seasonal_coefs1 = as.numeric(seasonal_coefs[1, 2:5])
station11 = 1
start_ind11 = 1
end_ind11 = 31
type11 = "HDD"
MLSS_spacetime11 = MLSS_spacetime(residuals1, station11, start_ind11,
                                  end_ind11, type11, seasonal_coefs1)
station12 = 1
start_ind12 = 182
end_ind12 = 212
type12 = "CDD"
MLSS_spacetime12 = MLSS_spacetime(residuals1, station12, start_ind12,
                                  end_ind12, type12, seasonal_coefs1)

# examine results
MLSS_spacetime11$price # HDD price
MLSS_spacetime12$price # CDD price
MLSS_spacetime11$mu # parameter mu of NIG random field
MLSS_spacetime12$mu # parameter mu of NIG random field

# example 2
residuals2 = matrix(as.numeric(residuals[, 51:55]), 730, 5)
seasonal_coefs2 = as.numeric(seasonal_coefs[51, 2:5])
station21 = 3
start_ind21 = 1
end_ind21 = 31
type21 = "HDD"
MLSS_spacetime21 = MLSS_spacetime(residuals2, station21, start_ind21,
                                  end_ind21, type21, seasonal_coefs2)
station22 = 3
start_ind22 = 182
end_ind22 = 212
type22 = "CAT"
MLSS_spacetime22 = MLSS_spacetime(residuals2, station22, start_ind22,
                                  end_ind22, type22, seasonal_coefs2)

# examine results
MLSS_spacetime21$price # HDD price
MLSS_spacetime22$price # CAT price
MLSS_spacetime21$mu # parameter mu of NIG random field
MLSS_spacetime22$mu # parameter mu of NIG random field
}
