% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\docType{data}
\name{residuals}
\alias{residuals}
\title{Historical deseasonalized temperatures}
\format{
A matrix with 730 rows and 227 columns
}
\usage{
residuals
}
\description{
A dataset containing the daily deseasonalized temperatures of
the 225 stations from 2020 to 2021
}
\keyword{datasets}
