% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_coefs.R
\docType{data}
\name{seasonal_coefs}
\alias{seasonal_coefs}
\title{Coefficients of the seasonal functions}
\format{
A matrix with 225 rows and 5 columns
}
\usage{
seasonal_coefs
}
\description{
A dataset containing the 4 coefficients of the fitted seasonal functions
for each of the 225 stations
}
\keyword{datasets}
