% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sresids.R
\docType{data}
\name{sresids}
\alias{sresids}
\title{Historical standardized residuals}
\format{
A matrix with 730 rows and 227 columns
}
\usage{
sresids
}
\description{
A dataset containing the daily standardized residuals of
the 225 stations from 2020 to 2021 after fitting the AR models
}
\keyword{datasets}
