% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp0.R
\docType{data}
\name{temp0}
\alias{temp0}
\title{Historical daily mean temperatures}
\format{
A matrix with 1,825 rows and 227 columns
}
\source{
National Centers for Environmental Information and World Meteorological Organization
}
\usage{
temp0
}
\description{
A dataset containing the daily mean temperatures of the 225 stations
from 2017 to 2021
}
\keyword{datasets}
