% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_forecast.R
\name{temp_forecast}
\alias{temp_forecast}
\title{Forecasting future temperatures}
\usage{
temp_forecast(residuals, station, seasonal_coefs, temp)
}
\arguments{
\item{residuals}{A n by p matrix containing the past deseasonalized temperatures data at p stations over (n / 365) years (n >= 365 * 2) (p >= 3)}

\item{station}{A numeric index denoting the specific station where forecasts are based on}

\item{seasonal_coefs}{A 1 by 4 vector containing the coefficients of the seasonality function at the station considered}

\item{temp}{A m by 1 vector containing the past temperatures data at one station over (m / 365) years}
}
\value{
A list containing:
\item{point}{a 1 by 365 vector containing the forecasted temperatures for the next year}
\item{lower}{a 1 by 365 vector containing the left endpoints of the confidence intervals for the forecasted temperatures}
\item{upper}{a 1 by 365 vector containing the right endpoints of the confidence intervals for the forecasted temperatures}
\item{newtemps}{a 1 by (m + 365) vector containing the past and future temperatures at that particular station}
}
\description{
Function that returns the point estimates and confidence intervals of the forecasted daily temperatures in the next year
after fitting the historical data to the spatio-temporal models
}
\examples{
# load the residuals data from residuals.rda in the data folder
# load the seasonal coefficients data from seasonal_coefs.rda in the data folder
# load the temp0 data from temp0.rda in the data folder

# example 1
residuals1 = matrix(as.numeric(residuals[, 3:5]), 730, 3)
seasonal_coefs1 = as.numeric(seasonal_coefs[1, 2:5])
temp = as.numeric(temp0[, 3]) # historical temperatures at one station
station11 = 1
temp_f1 = temp_forecast(residuals1, station11, seasonal_coefs1, temp)

# examine results
temp_f1$point # point estimates
temp_f1$lower # left endpoints of confidence intervals
temp_f1$upper # right endpoints of confidence intervals
temp_f1$newtemps # past plus futures forecasted temperatures
length(temp_f1$newtemps) # 365 * 6 = 2190

# example 2
residuals2 = matrix(as.numeric(residuals[, 51:55]), 730, 5)
seasonal_coefs2 = as.numeric(seasonal_coefs[51, 2:5])
temp = as.numeric(temp0[, 53]) # historical temperatures at one station
station21 = 3
temp_f2 = temp_forecast(residuals2, station21, seasonal_coefs2, temp)

# examine results
temp_f2$point # point estimates
temp_f2$lower # left endpoints of confidence intervals
temp_f2$upper # right endpoints of confidence intervals
temp_f2$newtemps # past plus futures forecasted temperatures
length(temp_f2$newtemps) # 365 * 6 = 2190
}
