% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempstations.R
\docType{data}
\name{tempstations}
\alias{tempstations}
\title{Stations}
\format{
A matrix with 225 rows and 7 columns
\describe {
\item{id}{id of each station}
\item{name}{name of the location}
\item{country}{country of the station, either US or CA}
\item{region}{the state of the station}
\item{latitude}{latitude of the station}
\item{longitude}{longitude of the station}
\item{altitude}{altitude of the station in meters}
}
}
\source{
National Centers for Environmental Information and World Meteorological Organization
}
\usage{
tempstations
}
\description{
A dataset containing the information of the 225 stations
}
\keyword{datasets}
