% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptBW.R
\name{adaptBW}
\alias{adaptBW}
\title{Pricing with adaptive bandwidth selection}
\usage{
adaptBW(residuals, station, start_ind, end_ind, type)
}
\arguments{
\item{residuals}{A n by p matrix containing the past deseasonalized temperatures data at p stations over (n / 365) years (n >= 365 * 2) (p >= 3)}

\item{station}{An numeric index denoting the specific station that the price of the temperature futures depends on}

\item{start_ind}{An numeric index for the start of the measurement period}

\item{end_ind}{An numeric index for the end of the measurement period}

\item{type}{A string denoting the type of the temperature futures, either CDD, HDD or CAT}
}
\value{
A list containing:
\item{price}{}
\item{sresids}{}
\item{plt}{}
}
\description{
Function that returns the price of a temperature futures using the adaptive bandwidth selection
method to estimate seasonal volatility
}
\examples{
# load the residuals data from residuals.rda in the data folder
residuals = matrix(as.numeric(residuals[, 3:5]), 730, 3)
}
